installed cors and added middleware

+++ MUNYUN/backend/controllers/expense.controller.js
@@ -1,4 +1,5 @@
 import Expense from '../models/expense.model.js';
+import mongoose from 'mongoose';
 
 export const getExpenses = async (req, res) => { 
     try {
@@ -48,12 +49,16 @@ export const updateExpense = async (req, res) => {
 
 export const deleteExpense = async (req, res) => {
     const {id} = req.params 
+
+    if ( !mongoose.Types.ObjectId.isValid(id)) { 
+        return res.status(404).json({ success: false, message: "invalid expense id"});
+    }
     
     try {
         await Expense.findByIdAndDelete(id)
         res.status(200).json({ success: true, message: 'Expense deleted successfully'})
     } catch (error) {
         console.log("error in deleting expense:", error.message);
-        res.status(404).json({ success: false, message: 'Expense not found' })
+        res.status(500).json({ success: false, message: 'Server Error' })
     }
 };
